Make subsubmodules private except base and canonical
