+++ src/prfmodel/density/__init__.py
@@ -1,8 +1,8 @@
 """Density functions."""
 
-from .gamma import derivative_gamma_density
-from .gamma import gamma_density
-from .gamma import shifted_gamma_density
+from ._gamma import derivative_gamma_density
+from ._gamma import gamma_density
+from ._gamma import shifted_gamma_density
 
 __all__ = [
     "derivative_gamma_density",
