+++ src/prfmodel/density/_gamma.py
