+++ src/prfmodel/fitters/__init__.py
@@ -1,11 +1,17 @@
 """Model fitters."""
 
-from .grid import GridFitter
-from .least_squares import LeastSquaresFitter
-from .sgd import SGDFitter
+from ._grid import GridFitter
+from ._grid import GridHistory
+from ._least_squares import LeastSquaresFitter
+from ._least_squares import LeastSquaresHistory
+from ._sgd import SGDFitter
+from ._sgd import SGDHistory
 
 __all__ = [
     "GridFitter",
+    "GridHistory",
     "LeastSquaresFitter",
+    "LeastSquaresHistory",
     "SGDFitter",
+    "SGDHistory",
 ]
