+++ src/prfmodel/fitters/_grid.py
