+++ src/prfmodel/fitters/_least_squares.py
