+++ src/prfmodel/fitters/_sgd.py
