+++ src/prfmodel/impulse/__init__.py
@@ -1,9 +1,9 @@
 """Impulse response models."""
 
-from .convolve import convolve_prf_impulse_response
-from .shifted_gamma import ShiftedGammaImpulse
-from .two_gamma import TwoGammaImpulse
-from .two_gamma_deriv import DerivativeTwoGammaImpulse
+from ._convolve import convolve_prf_impulse_response
+from ._shifted_gamma import ShiftedGammaImpulse
+from ._two_gamma import TwoGammaImpulse
+from ._two_gamma_deriv import DerivativeTwoGammaImpulse
 
 __all__ = [
     "DerivativeTwoGammaImpulse",
