+++ src/prfmodel/impulse/_convolve.py
