+++ src/prfmodel/impulse/_two_gamma.py
@@ -1,10 +1,9 @@
-"""Weighted difference of two derivative gamma distribution impulse response."""
+"""Weighted difference of two gamma distribution impulse response."""
 
 import pandas as pd
 from keras import ops
 from prfmodel._docstring import doc
-from prfmodel.density.gamma import derivative_gamma_density
-from prfmodel.density.gamma import gamma_density
+from prfmodel.density._gamma import gamma_density
 from prfmodel.typing import Tensor
 from prfmodel.utils import convert_parameters_to_tensor
 from prfmodel.utils import get_dtype
@@ -12,16 +11,14 @@ from prfmodel.utils import normalize_response
 from .base import BaseImpulse
 
 
-class DerivativeTwoGammaImpulse(BaseImpulse):
+class TwoGammaImpulse(BaseImpulse):
     r"""
-    Weighted difference of two derivative gamma distributions impulse response model.
+    Weighted difference of two gamma distributions impulse response model.
 
-    Predicts an impulse response that is the weighted derivative difference of two gamma distributions. This
-    weighted derivative difference is added to the weighted difference of the two gamma distributions.
-    The model has six parameters: `delay` and `undershoot` refer to the positive and negative peaks of the response
+    Predicts an impulse response that is the weighted difference of two gamma distributions.
+    The model has five parameters: `delay` and `undershoot` refer to the positive and negative peaks of the response
     while `dispersion` and `u_dispersion` refer to the rate parameters of the two gamma distributions. The `ratio`
-    parameter indicates the weight of the second gamma distribution. The `weight_deriv` represents the weight of the
-    derivative difference added to the standard difference.
+    parameter indicates the weight of the second gamma distribution.
 
     Parameters
     ----------
@@ -41,26 +38,17 @@ class DerivativeTwoGammaImpulse(BaseImpulse):
 
     See Also
     --------
-    TwoGammaImpulse : Weighted difference of two gamma distributions impulse response model.
     gamma_density : Density of the gamma distribution.
-    derivative_gamma_density : Derivative density of the gamma distribution.
 
     Notes
     -----
     The predicted impulse response at time :math:`t` with :math:`\alpha_1 = delay / dispersion`,
     :math:`\lambda_1 = dispersion`, :math:`\alpha_2  = undershoot / u\_dispersion`, :math:`\lambda_2 = u\_dispersion`,
-    :math:`\omega = ratio`, and :math:`\tau = weight\_deriv` is:
+    and :math:`\omega = ratio` is:
 
     .. math::
 
-        f(t) = f_{\text{diff}}(t) - \tau f'_{\text{diff}}(t)
-
-    .. math::
-
-        f_{\text{diff}}(t) = f_{\text{gamma}}(t; \alpha_1, \lambda_1) - \omega
-            f_{\text{gamma}}(t; \alpha_2, \lambda_2)
-
-    Positive `weight_deriv` values shift the response to the right.
+        f(t) = f_{\text{gamma}}(t; \alpha_1, \lambda_1) - \omega f_{\text{gamma}}(t; \alpha_2, \lambda_2)
 
     References
     ----------
@@ -72,6 +60,7 @@ class DerivativeTwoGammaImpulse(BaseImpulse):
     .. [3] Glover, G. H. (1999). Deconvolution of impulse response in event-related BOLD fMRI. *NeuroImage*, 9(4),
         416-429. https://doi.org/10.1006/nimg.1998.0419
 
+
     Examples
     --------
     >>> import pandas as pd
@@ -81,9 +70,8 @@ class DerivativeTwoGammaImpulse(BaseImpulse):
     ...     "undershoot": [1.5, 2.0, 1.0],
     ...     "u_dispersion": [1.0, 1.0, 1.0],
     ...     "ratio": [0.7, 0.2, 0.5],
-    ...     "weight_deriv": [0.5, -0.7, 0.9],
     ... })
-    >>> impulse_model = DerivativeTwoGammaImpulse(
+    >>> impulse_model = TwoGammaImpulse(
     ...     duration=100.0  # 100 seconds
     ... )
     >>> resp = impulse_model(params)
@@ -97,10 +85,10 @@ class DerivativeTwoGammaImpulse(BaseImpulse):
         """
         Names of parameters used by the model.
 
-        Parameter names are: `delay`, `dispersion`, `undershoot`, `u_dispersion`, `ratio`, `weight_deriv`.
+        Parameter names are: `delay`, `dispersion`, `undershoot`, `u_dispersion`, `ratio`.
 
         """
-        return ["delay", "dispersion", "undershoot", "u_dispersion", "ratio", "weight_deriv"]
+        return ["delay", "dispersion", "undershoot", "u_dispersion", "ratio"]
 
     @doc
     def __call__(self, parameters: pd.DataFrame, dtype: str | None = None) -> Tensor:
@@ -127,15 +115,8 @@ class DerivativeTwoGammaImpulse(BaseImpulse):
         undershoot = convert_parameters_to_tensor(parameters[["undershoot"]], dtype=dtype)
         u_dispersion = convert_parameters_to_tensor(parameters[["u_dispersion"]], dtype=dtype)
         ratio = convert_parameters_to_tensor(parameters[["ratio"]], dtype=dtype)
-        weight_deriv = convert_parameters_to_tensor(parameters[["weight_deriv"]], dtype=dtype)
 
         dens_1 = gamma_density(frames, delay / dispersion, dispersion)
         dens_2 = gamma_density(frames, undershoot / u_dispersion, u_dispersion)
 
-        dens_deriv_1 = derivative_gamma_density(frames, delay / dispersion, dispersion)
-        dens_deriv_2 = derivative_gamma_density(frames, undershoot / u_dispersion, u_dispersion)
-
-        diff_dens = dens_1 - ratio * dens_2
-        diff_dens_deriv = dens_deriv_1 - ratio * dens_deriv_2
-
-        return normalize_response(diff_dens - weight_deriv * diff_dens_deriv, self.norm)
+        return normalize_response(dens_1 - ratio * dens_2, self.norm)
