+++ src/prfmodel/impulse/_two_gamma_deriv.py
@@ -1,9 +1,10 @@
-"""Shifted gamma distribution impulse response."""
+"""Weighted difference of two derivative gamma distribution impulse response."""
 
 import pandas as pd
 from keras import ops
 from prfmodel._docstring import doc
-from prfmodel.density.gamma import shifted_gamma_density
+from prfmodel.density._gamma import derivative_gamma_density
+from prfmodel.density._gamma import gamma_density
 from prfmodel.typing import Tensor
 from prfmodel.utils import convert_parameters_to_tensor
 from prfmodel.utils import get_dtype
@@ -11,13 +12,16 @@ from prfmodel.utils import normalize_response
 from .base import BaseImpulse
 
 
-class ShiftedGammaImpulse(BaseImpulse):
+class DerivativeTwoGammaImpulse(BaseImpulse):
     r"""
-    Shifted gamma distribution impulse response model.
+    Weighted difference of two derivative gamma distributions impulse response model.
 
-    Predicts an impulse response that is a shifted gamma distribution.
-    The model has three parameters: `delay` refers to the positive peak, `dispersion` to the
-    rate, and `shift` to the onset of the gamma distribution.
+    Predicts an impulse response that is the weighted derivative difference of two gamma distributions. This
+    weighted derivative difference is added to the weighted difference of the two gamma distributions.
+    The model has six parameters: `delay` and `undershoot` refer to the positive and negative peaks of the response
+    while `dispersion` and `u_dispersion` refer to the rate parameters of the two gamma distributions. The `ratio`
+    parameter indicates the weight of the second gamma distribution. The `weight_deriv` represents the weight of the
+    derivative difference added to the standard difference.
 
     Parameters
     ----------
@@ -37,19 +41,26 @@ class ShiftedGammaImpulse(BaseImpulse):
 
     See Also
     --------
+    TwoGammaImpulse : Weighted difference of two gamma distributions impulse response model.
     gamma_density : Density of the gamma distribution.
-    shifted_gamma_density : Shifted density of the gamma distribution.
+    derivative_gamma_density : Derivative density of the gamma distribution.
 
     Notes
     -----
-    The predicted impulse response at time :math:`t` with :math:`\alpha = delay / dispersion`,
-    :math:`\lambda = dispersion`, and :math:`\delta = shift` is:
+    The predicted impulse response at time :math:`t` with :math:`\alpha_1 = delay / dispersion`,
+    :math:`\lambda_1 = dispersion`, :math:`\alpha_2  = undershoot / u\_dispersion`, :math:`\lambda_2 = u\_dispersion`,
+    :math:`\omega = ratio`, and :math:`\tau = weight\_deriv` is:
 
     .. math::
 
-        f(t) = f_{\text{gamma}}(t - \delta; \alpha, \lambda)
+        f(t) = f_{\text{diff}}(t) - \tau f'_{\text{diff}}(t)
 
-    The response prior to the onset of the gamma distribution is set to zero.
+    .. math::
+
+        f_{\text{diff}}(t) = f_{\text{gamma}}(t; \alpha_1, \lambda_1) - \omega
+            f_{\text{gamma}}(t; \alpha_2, \lambda_2)
+
+    Positive `weight_deriv` values shift the response to the right.
 
     References
     ----------
@@ -67,9 +78,12 @@ class ShiftedGammaImpulse(BaseImpulse):
     >>> params = pd.DataFrame({
     ...     "delay": [2.0, 1.0, 1.5],
     ...     "dispersion": [1.0, 1.0, 1.0],
-    ...     "shift": [1.0, 2.0, 5.0],
+    ...     "undershoot": [1.5, 2.0, 1.0],
+    ...     "u_dispersion": [1.0, 1.0, 1.0],
+    ...     "ratio": [0.7, 0.2, 0.5],
+    ...     "weight_deriv": [0.5, -0.7, 0.9],
     ... })
-    >>> impulse_model = ShiftedGammaImpulse(
+    >>> impulse_model = DerivativeTwoGammaImpulse(
     ...     duration=100.0  # 100 seconds
     ... )
     >>> resp = impulse_model(params)
@@ -83,10 +97,10 @@ class ShiftedGammaImpulse(BaseImpulse):
         """
         Names of parameters used by the model.
 
-        Parameter names are: `delay`, `dispersion`, and `shift`.
+        Parameter names are: `delay`, `dispersion`, `undershoot`, `u_dispersion`, `ratio`, `weight_deriv`.
 
         """
-        return ["delay", "dispersion", "shift"]
+        return ["delay", "dispersion", "undershoot", "u_dispersion", "ratio", "weight_deriv"]
 
     @doc
     def __call__(self, parameters: pd.DataFrame, dtype: str | None = None) -> Tensor:
@@ -107,10 +121,21 @@ class ShiftedGammaImpulse(BaseImpulse):
         parameters = self._join_default_parameters(parameters)
         dtype = get_dtype(dtype)
         frames = ops.cast(self.frames, dtype=dtype)
+
         delay = convert_parameters_to_tensor(parameters[["delay"]], dtype=dtype)
         dispersion = convert_parameters_to_tensor(parameters[["dispersion"]], dtype=dtype)
-        shift = convert_parameters_to_tensor(parameters[["shift"]], dtype=dtype)
+        undershoot = convert_parameters_to_tensor(parameters[["undershoot"]], dtype=dtype)
+        u_dispersion = convert_parameters_to_tensor(parameters[["u_dispersion"]], dtype=dtype)
+        ratio = convert_parameters_to_tensor(parameters[["ratio"]], dtype=dtype)
+        weight_deriv = convert_parameters_to_tensor(parameters[["weight_deriv"]], dtype=dtype)
 
-        dens = shifted_gamma_density(frames, delay / dispersion, dispersion, shift)
+        dens_1 = gamma_density(frames, delay / dispersion, dispersion)
+        dens_2 = gamma_density(frames, undershoot / u_dispersion, u_dispersion)
 
-        return normalize_response(dens, self.norm)
+        dens_deriv_1 = derivative_gamma_density(frames, delay / dispersion, dispersion)
+        dens_deriv_2 = derivative_gamma_density(frames, undershoot / u_dispersion, u_dispersion)
+
+        diff_dens = dens_1 - ratio * dens_2
+        diff_dens_deriv = dens_deriv_1 - ratio * dens_deriv_2
+
+        return normalize_response(diff_dens - weight_deriv * diff_dens_deriv, self.norm)
