+++ src/prfmodel/models/cf/__init__.py
@@ -1,13 +1,11 @@
 """Connective field models."""
 
-from .canonical import CanonicalCFModel
-from .gaussian import GaussianCFModel
-from .gaussian import GaussianCFResponse
-from .stimulus_encoding import CFStimulusEncoder
+from ._gaussian import GaussianCFModel
+from ._gaussian import GaussianCFResponse
+from ._stimulus_encoding import CFStimulusEncoder
 
 __all__ = [
     "CFStimulusEncoder",
-    "CanonicalCFModel",
     "GaussianCFModel",
     "GaussianCFResponse",
 ]
