+++ src/prfmodel/models/cf/_stimulus_encoding.py
