+++ src/prfmodel/models/prf/_stimulus_encoding.py
