+++ src/prfmodel/scaling/__init__.py
@@ -1,8 +1,8 @@
 """Scaling models."""
 
-from .amplitude import BaselineAmplitude
-from .div_norm_amplitude import DivNormAmplitude
-from .dog_amplitude import DoGAmplitude
+from ._amplitude import BaselineAmplitude
+from ._div_norm_amplitude import DivNormAmplitude
+from ._dog_amplitude import DoGAmplitude
 
 __all__ = [
     "BaselineAmplitude",
