+++ src/prfmodel/scaling/_amplitude.py
