+++ src/prfmodel/scaling/_div_norm_amplitude.py
