+++ src/prfmodel/scaling/_dog_amplitude.py
