+++ src/prfmodel/stimuli/_cf.py
