+++ src/prfmodel/stimuli/_prf.py
