+++ tests/impulse/test_convolve.py
@@ -3,9 +3,9 @@
 import numpy as np
 import pytest
 from prfmodel.exceptions import BatchDimensionError
-from prfmodel.impulse.convolve import _pad_response
-from prfmodel.impulse.convolve import _prepare_prf_impulse_response
-from prfmodel.impulse.convolve import convolve_prf_impulse_response
+from prfmodel.impulse import convolve_prf_impulse_response
+from prfmodel.impulse._convolve import _pad_response
+from prfmodel.impulse._convolve import _prepare_prf_impulse_response
 
 
 def test_pad_response():
