+++ tests/models/cf/test_css.py
@@ -3,10 +3,10 @@
 import pandas as pd
 import pytest
 from pytest_regressions.num_regression import NumericRegressionFixture
-from prfmodel.models.cf.gaussian import GaussianCFModel
-from prfmodel.models.cf.stimulus_encoding import CFStimulusEncoder
+from prfmodel.models.cf import CFStimulusEncoder
+from prfmodel.models.cf import GaussianCFModel
 from prfmodel.models.compression import CompressiveEncoder
-from prfmodel.stimuli.cf import CFStimulus
+from prfmodel.stimuli import CFStimulus
 from tests.models.conftest import CFSetup
