+++ tests/models/cf/test_encoding.py
@@ -3,9 +3,9 @@
 import numpy as np
 import pandas as pd
 import pytest
-from prfmodel.models.cf.stimulus_encoding import CFStimulusEncoder
+from prfmodel.models.cf import CFStimulusEncoder
 from prfmodel.models.compression import CompressiveEncoder
-from prfmodel.stimuli.cf import CFStimulus
+from prfmodel.stimuli import CFStimulus
 from tests.models.conftest import parametrize_dtype
