+++ tests/models/prf/test_gaussian.py
@@ -10,17 +10,17 @@ from prfmodel.exceptions import ShapeError
 from prfmodel.impulse import DerivativeTwoGammaImpulse
 from prfmodel.impulse.base import BaseImpulse
 from prfmodel.models.base import BaseEncoder
-from prfmodel.models.prf.gaussian import Gaussian2DPRFModel
-from prfmodel.models.prf.gaussian import Gaussian2DPRFResponse
-from prfmodel.models.prf.gaussian import GridMuDimensionsError
-from prfmodel.models.prf.gaussian import _check_gaussian_args
-from prfmodel.models.prf.gaussian import _expand_gaussian_args
-from prfmodel.models.prf.gaussian import predict_gaussian_response
-from prfmodel.models.prf.stimulus_encoding import PRFStimulusEncoder
+from prfmodel.models.prf import Gaussian2DPRFModel
+from prfmodel.models.prf import Gaussian2DPRFResponse
+from prfmodel.models.prf import PRFStimulusEncoder
+from prfmodel.models.prf import predict_gaussian_response
+from prfmodel.models.prf._gaussian import GridMuDimensionsError
+from prfmodel.models.prf._gaussian import _check_gaussian_args
+from prfmodel.models.prf._gaussian import _expand_gaussian_args
 from prfmodel.scaling import BaselineAmplitude
 from prfmodel.scaling.base import BaseScaling
-from prfmodel.stimuli.prf import GridDimensionsError
-from prfmodel.stimuli.prf import PRFStimulus
+from prfmodel.stimuli import GridDimensionsError
+from prfmodel.stimuli import PRFStimulus
 from tests.conftest import PRFStimulusSetup
 from tests.models.conftest import PRFStimulusGridSetup
 from tests.models.conftest import parametrize_dtype
